Upgraded to netstandard2.0 and net452, enabled NuGet package creation and refactored tests to use MSTestV2
